2019_hands-on-automated-machine-learning/py3-requirements.txt
# System requirement:
#   sudo apt install swig

auto-sklearn
TPOT
